GPS Brainstorming Technique
A brainstorming helps to quickly obtain new ideas and projects

The GPS brainstormkit is a structured method to brainstorm with a group of 12 to 15 people. We do that based on six trends and developments where the organization stands today. In half a day we come to 120 ideas and we distill from them three projects worked out where we can get to work. This GPS is like that in the car: be sure which way to go, then GPS indicates your organization the way!